+++ sdk/core/azure-core/src/http/raw_response.cpp
@@ -18,38 +18,6 @@ std::string const& RawResponse::GetReasonPhrase() const { return m_reasonPhrase;
 
 Azure::Core::CaseInsensitiveMap const& RawResponse::GetHeaders() const { return this->m_headers; }
 
-void RawResponse::SetHeader(uint8_t const* const first, uint8_t const* const last)
-{
-  // get name and value from header
-  auto start = first;
-  auto end = std::find(start, last, ':');
-
-  if (end == last)
-  {
-    throw std::invalid_argument("Invalid header. No delimiter ':' found.");
-  }
-
-  // Always toLower() headers
-  auto headerName = Azure::Core::_internal::StringExtensions::ToLower(std::string(start, end));
-  start = end + 1; // start value
-  while (start < last && (*start == ' ' || *start == '\t'))
-  {
-    ++start;
-  }
-
-  end = std::find(start, last, '\r');
-  auto headerValue = std::string(start, end); // remove \r
-
-  SetHeader(headerName, headerValue);
-}
-
-void RawResponse::SetHeader(std::string const& header)
-{
-  return SetHeader(
-      reinterpret_cast<uint8_t const*>(header.data()),
-      reinterpret_cast<uint8_t const*>(header.data() + header.size()));
-}
-
 void RawResponse::SetHeader(std::string const& name, std::string const& value)
 {
   return _detail::InsertHeaderWithValidation(this->m_headers, name, value);