+++ sdk/core/azure-core/src/http/request.cpp
@@ -76,15 +76,15 @@ std::string Request::GetHeadersAsString() const
 
 // Writes an HTTP request with RFC 7230 without the body (head line and headers)
 // https://tools.ietf.org/html/rfc7230#section-3.1.1
-std::string Request::GetHTTPMessagePreBody() const
-{
-  std::string httpRequest(HttpMethodToString(this->m_method));
-  // HTTP version harcoded to 1.0
-  auto const url = this->m_url.GetRelativeUrl();
-  httpRequest += " /" + url + " HTTP/1.1\r\n";
+// std::string Request::GetHTTPMessagePreBody() const
+// {
+//   std::string httpRequest(HttpMethodToString(this->m_method));
+//   // HTTP version harcoded to 1.0
+//   auto const url = this->m_url.GetRelativeUrl();
+//   httpRequest += " /" + url + " HTTP/1.1\r\n";
 
-  // headers
-  httpRequest += GetHeadersAsString();
+//   // headers
+//   httpRequest += GetHeadersAsString();
 
-  return httpRequest;
-}
+//   return httpRequest;
+// }