+++ sdk/core/azure-core/test/ut/transport_adapter_base.cpp
@@ -329,7 +329,6 @@ namespace Azure { namespace Core { namespace Test {
     auto request
         = Azure::Core::Http::Request(Azure::Core::Http::HttpMethod::Put, host, &bodyRequest);
     // Make transport adapter to read all stream content for uploading instead of chunks
-    request.SetUploadChunkSize(1024 * 1024);
     {
       auto response = m_pipeline->Send(request, Azure::Core::Context::GetApplicationContext());
       checkResponseCode(response->GetStatusCode());
@@ -449,8 +448,6 @@ namespace Azure { namespace Core { namespace Test {
     Azure::Core::IO::FileBodyStream requestBodyStream(testDataPath);
     auto request = Azure::Core::Http::Request(
         Azure::Core::Http::HttpMethod::Put, host, &requestBodyStream, true);
-    // Make transport adapter to read all stream content for uploading instead of chunks
-    request.SetUploadChunkSize(Azure::Core::Test::Datails::FileSize);
     {
       auto response = m_pipeline->Send(request, Azure::Core::Context::GetApplicationContext());
       checkResponseCode(response->GetStatusCode());
@@ -502,8 +499,6 @@ namespace Azure { namespace Core { namespace Test {
     Azure::Core::IO::FileBodyStream requestBodyStream(testDataPath);
     auto request = Azure::Core::Http::Request(
         Azure::Core::Http::HttpMethod::Put, host, &requestBodyStream, true);
-    // Make transport adapter to read more than file size (5Mb)
-    request.SetUploadChunkSize(Azure::Core::Test::Datails::FileSize * 5);
     {
       auto response = m_pipeline->Send(request, Azure::Core::Context::GetApplicationContext());
       checkResponseCode(response->GetStatusCode());